LayoutTests/js/dom/modules/module-execution-order-mixed-with-classic-scripts.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test module execution order between mixed tags.');
debug('We keep the execution order between modules.');
window.count = 0;

// Module will be executed asynchronously.
window.jsTestIsAsync = true;
debug('Module is not executed yet.');
</script>
<script src="../../../resources/js-test-post.js"></script>
<script>
shouldBe("count++", "0");
</script>
<script>
shouldBe("count++", "1");
</script>
<script type="module" src="resources/module-execution-order-mixed-with-classic-scripts.js"></script>
<script>
shouldBe("count++", "2");
</script>
<script type="module">
shouldBe("count++", "10");
</script>
<script>
shouldBe("count++", "3");
</script>
<script type="module" src="resources/module-execution-order-mixed-with-classic-scripts-2.js"></script>
<script>
shouldBe("count++", "4");
</script>
<script type="module">
shouldBe("count++", "12");
finishJSTest();
</script>
<script>
shouldBe("count++", "5");
</script>
</body>
</html>
