Safari-609.1.11.1/LayoutTests/js/dom/modules/module-execution-order-mixed-with-classic-scripts-expected.txt

Test module execution order between mixed tags.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


We keep the execution order between modules.
Module is not executed yet.
PASS count++ is 0
PASS count++ is 1
PASS count++ is 2
PASS count++ is 3
PASS count++ is 4
PASS count++ is 5
PASS count++ is 6
PASS count++ is 7
PASS count++ is 8
PASS count++ is 9
PASS count++ is 10
PASS count++ is 11
PASS count++ is 12
PASS successfullyParsed is true

TEST COMPLETE
